+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Below you will find all the different changes that have been added since the fir
 * Activity Indicator: Added new BlackBerry 10 Activity Indicator
 * CheckBoxes: Added new BlackBerry 10 styling to check boxes
 * Grid Layout: Added animation to images on orientation change for grid layouts
+* Scroll Panel: Added panel where all inner contents will scroll
+* Screen Transitions: Slide screen transitions are now supported for BlackBerry 10 and PlayBook 
 * Noteable Bug Fixes: 
     * Action Bar overflow used to appear briefly when changing orientation
     * Solid headers on image lists used to have a JavaScript exception